# perf(trie-router): avoid calling spread operator for `Object.create(null)`