--- a/src/router/trie-router/node.ts
+++ b/src/router/trie-router/node.ts
@@ -83,8 +83,8 @@ export class Node<T> {
   #getHandlerSets(
     node: Node<T>,
     method: string,
-    nodeParams: Record<string, string>,
-    params: Record<string, string>
+    nodeParams?: Record<string, string>,
+    params?: Record<string, string>
   ): HandlerParamsSet<T>[] {
     const handlerSets: HandlerParamsSet<T>[] = []
     for (let i = 0, len = node.#methods.length; i < len; i++) {
@@ -93,15 +93,16 @@ export class Node<T> {
       const processedSet: Record<number, boolean> = {}
       if (handlerSet !== undefined) {
         handlerSet.params = Object.create(null)
-        for (let i = 0, len = handlerSet.possibleKeys.length; i < len; i++) {
-          const key = handlerSet.possibleKeys[i]
-          const processed = processedSet[handlerSet.score]
-          handlerSet.params[key] =
-            params[key] && !processed ? params[key] : nodeParams[key] ?? params[key]
-          processedSet[handlerSet.score] = true
-        }
-
         handlerSets.push(handlerSet)
+        if (nodeParams) {
+          for (let i = 0, len = handlerSet.possibleKeys.length; i < len; i++) {
+            const key = handlerSet.possibleKeys[i]
+            const processed = processedSet[handlerSet.score]
+            handlerSet.params[key] =
+              params?.[key] && !processed ? params[key] : nodeParams[key] ?? params?.[key]
+            processedSet[handlerSet.score] = true
+          }
+        }
       }
     }
     return handlerSets
@@ -130,18 +131,9 @@ export class Node<T> {
           if (isLast) {
             // '/hello/*' => match '/hello'
             if (nextNode.#children['*']) {
-              handlerSets.push(
-                ...this.#getHandlerSets(
-                  nextNode.#children['*'],
-                  method,
-                  node.#params,
-                  Object.create(null)
-                )
-              )
+              handlerSets.push(...this.#getHandlerSets(nextNode.#children['*'], method))
             }
-            handlerSets.push(
-              ...this.#getHandlerSets(nextNode, method, node.#params, Object.create(null))
-            )
+            handlerSets.push(...this.#getHandlerSets(nextNode, method))
           } else {
             tempNodes.push(nextNode)
           }
@@ -156,9 +148,7 @@ export class Node<T> {
           if (pattern === '*') {
             const astNode = node.#children['*']
             if (astNode) {
-              handlerSets.push(
-                ...this.#getHandlerSets(astNode, method, node.#params, Object.create(null))
-              )
+              handlerSets.push(...this.#getHandlerSets(astNode, method, node.#params))
               tempNodes.push(astNode)
             }
             continue