This optimization is based on the discussion in the following Pull Request.
Thanks to @EdamAme-x!
https://github.com/honojs/hono/pull/3732


### benchmarks

via benchmarks/routers

#### node

```
• all together
----------------------------------------------------------------------- -----------------------------
Hono RegExpRouter                   513 ns/iter       (490 ns … 715 ns)    524 ns    595 ns    715 ns
Hono TrieRouter(perf-trie-1208)   1'716 ns/iter   (1'652 ns … 1'895 ns)  1'723 ns  1'892 ns  1'895 ns
Hono TrieRouter(main)             2'203 ns/iter   (2'097 ns … 2'730 ns)  2'213 ns  2'709 ns  2'730 ns

summary for all together
  Hono RegExpRouter
   3.35x faster than Hono TrieRouter(perf-trie-1208)
   4.3x faster than Hono TrieRouter(main)
```

#### bun

```
• all together
----------------------------------------------------------------------- -----------------------------
Hono RegExpRouter                   393 ns/iter       (360 ns … 722 ns)    390 ns    594 ns    722 ns
Hono TrieRouter(perf-trie-1208)   3'191 ns/iter   (2'032 ns … 3'602 ns)  3'351 ns  3'595 ns  3'602 ns
Hono TrieRouter(main)             3'971 ns/iter   (3'732 ns … 4'293 ns)  4'065 ns  4'267 ns  4'293 ns

summary for all together
  Hono RegExpRouter
   8.12x faster than Hono TrieRouter(perf-trie-1208)
   10.1x faster than Hono TrieRouter(main)
```


### The author should do the following, if applicable

- <del>Add tests</del>
- [x] Run tests
- [x] `bun run format:fix && bun run lint:fix` to format the code
- <del>Add [TSDoc](https://tsdoc.org/)/[JSDoc](https://jsdoc.app/about-getting-started) to document the code</del>