@usualoma 2024-12-08
There was a slight performance drop due to the [eac0056](https://github.com/honojs/hono/pull/3735/commits/eac0056bc0ae73408fb6da45dce047ed1e17ad72)  bug fix commit, but I still think it's the fastest.