@yusukebe 2024-12-09
I also like the not-magical way. This is simple and good! 

Thanks, @usualoma @EdamAme-x !